+++ components/CreditsItem.vue
@@ -38,7 +38,7 @@
 </template>
 
 <script>
-import { TMDB_IMAGE_URL } from '~/config/tmdb';
+import { TMDB_IMAGE_URL } from '~/config/tmdbAPI';
 
 export default {
   props: {
