+++ components/person/PersonInfo.vue
@@ -95,7 +95,7 @@
 </template>
 
 <script>
-import { TMDB_IMAGE_URL } from '~/config/tmdb';
+import { TMDB_IMAGE_URL } from '~/config/tmdbAPI';
 import ExternalLinks from '~/components/ExternalLinks';
 
 export default {
